add -no entries-Message and delete dummy-entries
