+++ lib/initialEntries.js
@@ -1,209 +1,3 @@
-const initialEntries = [
-  {
-    id: "1",
-    type: "PastAnalysis",
-    title: "Argument with friend",
-    date: "2022-10-01",
-    description: "I had an argument with my friend",
-    interpretations: [
-      { id: "1", interpretation: "My friend doesn't care about me" },
-      { id: "2", interpretation: "I always ruin my friendships" },
-      { id: "3", interpretation: "I can't trust anyone" },
-    ],
-    behavior: "I stopped talking to my friend",
-    behaviorKieslerkreis: "[null,null,null,null,2,null,null,null]",
-    actualResult: "We grew apart",
-    desiredResult: "To resolve the argument and maintain the friendship",
-    comparison: "The actual result was not what I desired",
-    revision: [
-      {
-        id: "1",
-        revision: "My friend may have been going through a tough time",
-      },
-      { id: "2", revision: "I may have overreacted in the situation" },
-      { id: "3", revision: "My friend may have had a different perspective" },
-    ],
-    actionInterpretation: "I am allowed to listen to my feelings",
-    behaviorChange: "Talk to my friend and try to resolve the argument",
-    behaviorChangeKieslerkreis: "[1,null,null,null,null,null,null,null]",
-    implementation: "I called my friend and we talked it out",
-    transfer: "I will try to communicate better in future disagreements",
-  },
-  {
-    id: "1test",
-    type: "PastAnalysis",
-    title: "TESTArgument with friend",
-    date: "2022-10-01",
-    behaviorKieslerkreis: "[null,null,3,null,null,null,null,null]",
-    description: "I had an argument with my friend",
-    interpretations: [
-      { id: "1", interpretation: "My friend doesn't care about me" },
-      { id: "2", interpretation: "I always ruin my friendships" },
-      { id: "3", interpretation: "I can't trust anyone" },
-    ],
-    behavior: "I stopped talking to my friend",
-    actualResult: "We grew apart",
-    desiredResult: "To resolve the argument and maintain the friendship",
-    comparison: "The actual result was not what I desired",
-    revision: "My interpretation may have been incorrect",
-    actionInterpretation: "My opinion has value",
-    behaviorChange: "Talk to my friend and try to resolve the argument",
-    behaviorChangeKieslerkreis: "[null,null,null,2,null,null,null,null]",
-    implementation: "I called my friend and we talked it out",
-    transfer: "I will try to communicate better in future disagreements",
-  },
-  {
-    id: "2",
-    type: "InnerSituationAnalysis",
-    title: "Anxiety in social situations",
-    date: "2022-11-15",
-    behaviorKieslerkreis: "[null,null,null,null,null,null,null,null]",
-    description: "I feel anxious in social situations",
-    interpretations: [
-      { id: "1", interpretation: "People are judging me" },
-      { id: "2", interpretation: "I don't fit in" },
-      { id: "3", interpretation: "I'm not interesting enough" },
-    ],
-    behavior: "I avoid social situations",
-    actualResult: "I feel isolated",
-    desiredResult: "To feel comfortable in social situations",
-    comparison: "The actual result was not what I desired",
-    revision: [
-      {
-        id: "1",
-        revision: "My friend may have been going through a tough time",
-      },
-      { id: "2", revision: "I may have overreacted in the situation" },
-      { id: "3", revision: "My friend may have had a different perspective" },
-    ],
-    actionInterpretation: "I am strong enough to overcome challenges",
-    behaviorChange: "Try to attend social situations and challenge my thoughts",
-    behaviorChangeKieslerkreis: "[null,null,null,null,null,null,null,null]",
-    implementation:
-      "I attended a social event and realized people were not judging me",
-    transfer: "I will try to challenge my thoughts in future social situations",
-  },
-  {
-    id: "3",
-    type: "PastAnalysis",
-    title: "Failed Exam",
-    behaviorKieslerkreis: "[null,null,null,null,null,null,null,null]",
-    date: "2022-09-01",
-    description: "I failed an important exam",
-    interpretations: [
-      { id: "1", interpretation: "I am not smart enough" },
-      { id: "2", interpretation: "I'm not cut out for this subject" },
-      { id: "3", interpretation: "Everyone else did better than me" },
-    ],
-    behavior: "I stopped studying for other exams",
-    actualResult: "My grades suffered",
-    desiredResult: "To pass the exam and improve my grades",
-    comparison: "The actual result was not what I desired",
-    revision: [
-      {
-        id: "1",
-        revision: "People may not be paying as much attention to me as I think",
-      },
-      {
-        id: "2",
-        revision: "I may have more in common with others than I realize",
-      },
-      { id: "3", revision: "I may have interesting things to share" },
-    ],
-    actionInterpretation: "I can learn and grow from my mistakes",
-    behaviorChange: "Study harder and ask for help if needed",
-    behaviorChangeKieslerkreis: "[null,null,null,null,null,null,null,null]",
-    implementation: "I studied harder and passed the next exam",
-    transfer: "I will continue to study hard and ask for help if needed",
-  },
-  {
-    id: "4",
-    type: "InnerSituationAnalysis",
-    title: "Fear of public speaking",
-    date: "2022-12-01",
-    behaviorKieslerkreis: "[null,null,null,3,null,null,null,null]",
-    description: "I am afraid of public speaking",
-    interpretations: [
-      {
-        id: "1",
-        interpretation: "People will laugh at me or think I'm stupid",
-      },
-      { id: "2", interpretation: "No one wants to hear what I have to say" },
-      { id: "3", interpretation: "I'll freeze up and forget everything" },
-    ],
-    behavior: "I avoid public speaking opportunities",
-    actualResult: "I miss out on opportunities to share my ideas",
-    desiredResult: "To feel confident while public speaking",
-    comparison: "The actual result was not what I desired",
-    revision: [
-      { id: "1", revision: "People may be interested in what I have to say" },
-      { id: "2", revision: "I may have valuable insights to share" },
-      {
-        id: "3",
-        revision: "I can prepare and practice to feel more confident",
-      },
-    ],
-    actionInterpretation: "",
-    behaviorChange: "Practice public speaking and challenge my thoughts",
-    behaviorChangeKieslerkreis: "[null,null,null,null,null,null,null,2]",
-    implementation:
-      "I practiced public speaking and realized people were interested in what I had to say",
-    transfer:
-      "I will continue to practice public speaking and challenge my thoughts",
-  },
-  {
-    id: "5",
-    type: "FutureAnalysis",
-    title: "Starting a new hobby",
-    date: "2023-02-01",
-    behaviorKieslerkreis: "[null,null,null,2,null,null,null,null]",
-    description: "I want to start a new hobby, painting",
-    desiredResult: "To enjoy painting and improve my skills over time",
-    behavior:
-      "Buy painting supplies and take a class or watch tutorials online",
-    actionInterpretation: [
-      {
-        id: "1",
-        actionInterpretation:
-          "Painting is something I can enjoy and improve at over time, even if I'm not good at it right away.",
-      },
-      {
-        id: "2",
-        actionInterpretation: "It's okay to make mistakes while learning",
-      },
-      {
-        id: "3",
-        actionInterpretation: "Art is about self-expression, not perfection",
-      },
-    ],
-    rolePlay: "",
-    implementation:
-      "I bought painting supplies, took a class, and started painting regularly. I enjoyed it and saw improvement over time.",
-    transfer:
-      "I will continue to paint regularly and enjoy the process of improving over time.",
-  },
-  {
-    id: "6",
-    type: "FutureAnalysis",
-    title: "Applying for a new job",
-    date: "2023-01-01",
-    description: "I want to apply for a new job",
-    desiredResult: "To get the job and feel fulfilled in my career",
-    behavior: "Prepare for the interview and present myself well",
-    behaviorKieslerkreis: "[2,null,null,null,null,null,null,null]",
-    actionInterpretation: [
-      {
-        id: "1",
-        actionInterpretation: "I am capable and qualified for this job",
-      },
-      { id: "2", actionInterpretation: "I have valuable skills to offer" },
-      { id: "3", actionInterpretation: "This is a good opportunity for me" },
-    ],
-    rolePlay: "Practiced interview questions with a friend",
-    implementation: "I attended the interview and felt confident",
-    transfer:
-      "I will use this experience to feel confident in future interviews",
-  },
-];
+const initialEntries = [];
 
 export default initialEntries;
